web/requirements.txt
tensorflow
flask
pillow
numpy
werkzeug
pyserial
